+++ CHANGELOG.md
@@ -6,11 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.0] - 2020-11-22
 ### Added
 * Support for cross compilation for `musl`
+* Check that run-dir and data-dir have correct rw permissions
+* runtime can be correctly shut down
 
 ### Changed
 * Rewrite of nstar for a better user experience
+* Simplified error types in runtime and API
+* Extract all npk handling to separate crate
+
+### Removed
+* Mock runtime support for MacOS
 
 ## [0.5.0] - 2020-11-2
 ### Added